Store parent interfaces only in ClassDef/ClassInfo.
Instead of parent class + parent interfaces in a field named
'parents', we now store only parent interfaces in a field named
'interfaces'. The superClass field also becomes an Option.

This is closer to how .class files are built, and is in general
cleaner. It differs from the .class file format by the fact that
we store None as superclass of interfaces, whereas in a .class
file the Object class is mentioned as superclass of any interface.
We choose None because classOf[SomeIntf].getSuperclass returns
null.