Add methods of j.l.{Byte,Short} to j.l.Integer.
When calling a method of j.l.Byte or j.l.Short on a primitive
value, it appears to be called directly on the primitive value,
which has type IntType. Call resolution, by the analyzer and the
optimizer, will then look for the method in the class j.l.Integer
instead of j.l.Byte or j.l.Short. This is why we add here the
methods of these two classes that are not already in j.l.Integer.