Factor out common code between ByteBuffers and CharBuffers.
To do this, we add in Buffer private[nio] generic accessors to
the fields and methods declared with specific types in the
subclasses of Buffer. These accessors connect directly with
the actual methods with bridges.

Using these accessors, we write generic algorithms, that are
private[nio] as well, for all the methods of {Byte,Char}Buffer
and Heap{Byte,Char}Buffer.

The whole thing would be totally inefficient because of all the
boxes and erasures of Array[T] to AnyRef, if it were not for
the optimizer. But in fact, it is completely capable of remove
all those indirections, and produce the same code as what we
had before the refactoring.