Intrinsify ScalaRunTime.array_{apply,update,length}.
These methods are used by the scalac codegen when it erases an
Array[T] to AnyRef, to be able to implement their methods apply,
update and length.

When, due to type refinement, the optimizer knows that such an
erased AnyRef is in fact a proper ArrayType of some sort, it can
emit a direct ArraySelect or ArrayLength node.