Compute reachability of instance tests and data separately.
This yields significantly smaller fastOpt outputs, but the
fullOpt ones seem to remain almost intact. This is expected,
since most of this stuff gets rid of *instance tests*, but
not so much *data*. And Closure is able to dce instance tests
easily, since they are top-level static functions.

This causes binary incompatibilities in the tools.