Optimize java.util.Random by not using Longs internally.
According to a user report, calling Math.random(), which delegates
to java.util.Random.nextDouble(), was a major bottleneck of an
algorithm, compared to the raw js.Math.random(), to the point
that it was unusably slow.

In this commit, we optimize java.util.Random.next(bits: Int),
as well nextDouble() and nextFloat(), essentially by *not* using
Longs. We still implement the specified algorithm (which is
defined in terms of Long operations) with clever bit and double
manipulations.