A new output mode that creates everything on the top-level.
Instead of namespacing everything in fields and subfields of
a single var ScalaJS, as in ScalaJS.imul or ScalaJS.c.Foo, this
output mode creates everything in top-level vars, such as
$imul or $c_Foo, respectively.

This mode is only safe if it is systematically wrapped in an
anonymous function, so it does that.

The benefit is that, even in fastOpt mode, we do not have the
overhead of the namespacing through objects.

This mode is not suitable to the way the Rhino interpreter works,
so the old mode is kept. A new OutputMode type determines how to
emit the final JavaScript code.

This OutputMode will find other uses in the future, for example
to target ECMAScript 6.