Revert "Fix #1299: Launch slaves in test adapter in anticipation"
This fixes #1556.

While in principle reducing latency for testing, the commit in
question has shown to launch too many concurrent slaves in some
cases. This happens especially if mutliple projects are tested at the
same time in sbt.

I do not see a way how we can determine the number of slaves we
provision in a way that interacts nicely with sbt's scheduling
(without going into very complicated heuristics). Since the old
adapter has shown to be good enough, we simply revert.

Changes to clean revert: Binary compatibiliy filter.

This reverts commit 6d61c347281b597af309bbf789f570feb5d90059.