Introduce ...rest parameters in the IR.
The magic `arguments` variable is now banned from the IR. It is
replaced by its cleaner ES6 replacement: ...rest parameters.

Rest params are only allowed in exported methods and constructors,
and only as last parameter. They are desugared into a loop
preceding the normal body of the method, copying the elements from
`arguments`. But until desugaring, we do not have to deal with it.

When eventually targeting ES6, a rest parameter need not be
desugared at all.

To preserve backward binary compatibility, a deserialization hack
identifies uses of `arguments`, and replaces it by a rest param.
All normal parameters are also removed, and replaced by indexed
access into the rest parameter.

This commit does not yet modify the output of the compiler, so that
we can test the deserialization hack.