Fix #1564: Allow filtering of JS dependency manifests
This allows to resolve relative path conflicts in manifests inherited
from dependencies as the test shows. The typical example (for which we
provide a convenience filter) is a dependency that does not declare
specific enough relative paths.

Binary compatibility of PartialClasspath.resolve is a bit funky here:
The default parameter for the one argument version is now provided by
the parameter of the two argument version. However, since the values
equal, this is fine. Also, argument names need not be deprecated, since
the overloading resolution works properly, if filter is used.