Fix #1581: Work around an unexpected ExistentialType in genReflCallProxy.
With scalac 2.10, when really complex types and methods are involved,
it may happen that the type of a method entering posterasure is an
ExistentialType, which should not happen. If it is, determining whether
and how to box it to generate the reflective call proxy crashes the
compiler.

This commit works around this bug by intercepting the ExistentialType
before it causes harm, and assuming that no boxing is required.