Fix #1586: Make Emitter agree with Optimizer on the id of a method.
For the optimizer, the identity of a method is its name + the list
of all classes in the hierarchy of its enclosing class.
Previously, the Emitter only used the enclosing class, but not its
hierarchy. This caused problems with the versioning of methods.

This commit solves the issue by making the Emitter agree with the
Optimizer that the identity of a method depends on the entire hierarchy.