Rewrite translated pattern matches as if/elses early in compiler.
This helps the optimizer a lot, because it can collapse if/else
chains much better than Labeled blocks. The optimizer already did
basically the same optimization, but only after it has processed
the body of the Labeled block, at which point it has lost
opportunities to alias LocalDefs and constant-fold.