Add runtime.linkingInfo to access link-time information.
This new primitive gives access to a global, constant object
with information available at link-time, depending on the
output configuration.

It contains everything the linker knows and that could influence
how things are emitted:

* Checked behaviors (asInstanceOfs and moduleInit)
* Whether floats have strict semantics
* Whether we assume that ECMAScript 6 is supported

The members of linkingInfo are constant-folded by the optimizer,
which allows to use these information efficiently, and, in
particular, to dead-code-eliminate anything that is not needed.

This is used to implement runtime.assumingES6, which was
previously an intrinsic.