Fix #1671: Fix {Float,Double}.toInt for large abs values.
The behavior of ToInt32(x) in ES does not match the behavior of
(int) x on the JVM for values that are lower than Int.MinValue
or larger than Int.MaxValue. On the JVM, and hence for .toInt,
they are clamped at Int.MinValue or Int.MaxValue, respectively.
So we have to handle these cases separately.