Fix #1647: Require explicit @JSExport("apply")
Exporting a method named apply (without explicit name) should cause it
to be callable using call synatx in JavaScript. This is not doable. We
therefore require explicitly specifying the name of the method to be
"apply". This will cause the method to callable under the name apply.

As a side effect, the clean up in this commit also fixes #1723 (export
object under the name toString).