Make the binary version of a minor snapshot be that of its release.
When working on 0.6.4-SNAPSHOT, for example, the binary version
should still be 0.6, and not 0.6.4-SNAPSHOT, since backward binary
compatibility is guaranteed wrt. 0.6.0. This is true even if the
version of the IR changes in a non-forward binary compatible way.

However, when working on 1.0.0-SNAPSHOT, there is no binary
compatibility guarantee wrt. 1.0.0, and hence the binary version
should be 1.0.0-SNAPSHOT.