Fix #1743: Accept expanded Seqs in js.Dynamic.literal().
Previously, a seq: _* passed as an argument to
js.Dynamic.literal.applyDynamic{,Named}
would cause the compiler to generate invalid IR, where a
JSSpread found its way in an invalid context.

This commit intercepts the cases where there is a JSSpread,
and calls a runtime method to turn the array of tuples into
an object.