Fix #1756: Use the linking info in ScalaJSClassEmitter.
The issue was caused by the following pattern:

class A
class B extends A
println(classOf[B])

In this case, `B` is retained by the linker because its runtime
type info are required. But `A` is discarded because it is not
used in any way, despite it being a superclass of `B`. This is
normal.

The problem is that the Rhino runner, when lazy loading `B`,
did not use the fact that `linkedClassB.hasInstances` is false,
and therefore generated the (empty) class. This empty class *did*
reference its parent class `A`, so when the script was evaluated,
it would look for `A`. However it could not find `A` at all
because it has been completely eliminated by the linker.

This commit fixes this by making ScalaJSClassEmitter leverage the
linking info available in the LinkedClass it prints to emit only
the necessary parts, basically with the same tests as those used
by Emitter.