Fix #1759: Use `new` when copying a TypedArray in Array.clone().
The ES6 spec says that it is illegal to call a TypedArray
constructor without `new`.

This does not come with any new test because we already have the
tests that would trigger this error. But we don't have any
testing runtime that triggers it: the latest versions of Node.js
and io.js are too lenient about this, and are fine with the
absence of `new`.