Refactor PrepJSInterop: clean up the owner chain handling.
PrepJSInterop performs various checks that vary according to the
"owner chain" of a class/trait/module or member definition.

Keeping track of where we are was done with a set of boolean fields
for various kinds of owner we wanted to track. This was hard to
read, and had the limitation that we could not tell information
about the directly enclosing owner, which made some checks even
less readable.

This commit refactors this, using a value type OwnerKind encoding
the precise kind of one owner, or a set of kinds we want to check
against. We maintain the precise kind of the directly enclosing
owner, as well as an accumulated set of all kinds of owners in the
current owner chain. This basically provides a DSL to easily and
readably check various conditions in a precise way.