Change the exception message used in UUID.fromString
The OpenJDKs 6 and 7 and the Oracle JDKs 7 and 8 use "Invalid UUID string"
as prefix for the exception message that is used in UUID.fromString. The
Android SDK's message begins with "Invalid UUID" but skips " string". This
changes the message from "Illegal UUID string" to "Invalid UUID string" to
match the message used by the aforementioned JDKs.