Fix #1840: Give a meaningful `name` property to emitted classes.
Instead of emitting
  var $c_pack_Foo = function() { ... };
we now emit
  function $c_pack_Foo() { ... }

Not only is this shorter, but most importantly it gives a meaningful
`name` property to the function. Some debugging facilities, including
those built in some frameworks, leverage this property.

Obviously, the name is only meaningful in fastOpt mode. In fullOpt,
the names are shortened anyway.