Fix #1820: Allow nested native JS stuff in native JS objects.
Native JS objects can now contain nested native JS traits, classes
and objects. For classes and objects, their full JS name is
qualified by the full JS name of the enclosing native JS object,
since they appear as members of this object. In this example:

object Foo extends js.Object {
  @JSName("Babar")
  object/class Bar extends js.Object
}

the full JS name of Bar is "Foo.Babar", not just "Babar".

This prompted to revisit native JS objects and classes in
Scala objects. For those, an explicit @JSName is required, and
is their full JS name (it is not qualified by their enclosing
Scala object). For classes, this is a warning, since we
previously erroneously allowed that.

Native JS stuff is not allowed in Scala.js-defined JS objects.

Altogether, these changes to the rules prompted a major refactoring
of the checks performed by PrepJSInterop.