Fix #1870: Add js.constructorOf[JSClass] to retrieve its constructor.
The method `js.constructorOf[JSClass]` returns the JavaScript
constructor function for the class `JSClass`. The latter must
be a class (not trait nor object) extending js.Any. It may be
a native JS class or a Scala.js-defined JS class. The restrictions
on what types can be used are enforced at compile-time.