Fix #1928: Defer calls to getSimpleName until needed in IR checker.
Instead of eagerly fetching the simple name of tree classes, we
just store the tree node itself. We only extract the position and
class name for `ErrorContext.toString()`, which is called only if
an error is reported. In successful runs, there is no error
reported, so we avoid a lot of allocations.

ErrorContext itself is also made an AnyVal to further reduce
allocations in the IR checker.