Fix #1955: Recognize an additional shape of Match in the back-end.
After mixin, scalac gives us the following tree for the bug()
method:

    def bug(e: Boolean): Unit = {
      {
        case <synthetic> val x1: Int = 1;
        (x1: Int) match {
          case 1 => Bug.this.doSomething(123, 456, scala.runtime.BoxedUnit.UNIT)
          case 2 => {
            if (e)
              ()
            else
              default4();
            scala.runtime.BoxedUnit.UNIT
          }
          case _ => default4(){
            throw new MatchError(scala.Int.box(x1))
          }
        }
      };
      ()
    };

in which the `case 2 =>` has the BoxedUnit.UNIT *after* the
if/else, instead of inside the branches, as one might expect.
We now recognize this shape, and rewrite it by injecting
`scala.runtime.BoxedUnit.UNIT` in the two branches of the if,
so that it fits the shape we can deal with.