Automatically generate Infos from the Trees in GenJSCode.
Previously, we used mutable builders, and we had to take care to
call the appropriate mutating methods when constructing immutable
Trees. This was extremely tedious, error-prone, and has been known
to be the cause of bugs in the past.

This commit completely removes all references to Infos from
GenJSCode and GenJSExports. Instead, we focus on building the
Trees. At the end of the process, GenJSFiles can automatically
generate the appropriate ClassInfo for the ClassDef that it needs
to serialize.

This considerably simplifies the code, at the cost of an
additional pass on the Trees to generate the Infos after the fact.