Fix #1963 and #2010: Prevent GCC to remove the `constructor` prop.
Under some circumstances, the Closure compiler would remove the
reassignment of the `constructor` property on the prototype of
classes. This removal caused the issues seen in #1963 and #2010.

This commit fixes this issue by assigning it as StringLiteral
instead of an Ident.

The tests for this fix are put in javalibExTestSuite instead of
testSuite because, if put in testSuite, they "fail to fail",
probably due to mass effects of the testSuite codebase, which
cause Closure not to exhibit the bug.