Fix #1843: Do not manually manage the statesInUse in OptimizerCore.
The fix for #1515 in 00bf9f9a25301349a70ca89a07b73cb8970b6eb0 had
to introduce a trampoline in withState, which basically defeated
most of the purpose of our trampolines, because each withNewLocalDef
had to introduce a new state.

This commit completely gets rid of withState. Instead of trying
(and mostly failing) to keep track of the list of states currently
in use, and remove the unused states, we simply keep them all
forever. The RollbackExceptions themselves keep track of which
states where in use when they were created, and restore only those.
In fact, they simply record `StateBackup`s, which keep their State
themselves.

This new scheme can create some amount of "memory leak", because
states that are not actually used anymore are kept in statesInUse,
and backups for them are made. However, that list is thrown out
with the OptimizerCore when `optimize` terminates anyway, so the
"leak" is bound in a short period of time.