Constructors are not inherited anymore at the IR level.
In JVM bytecode, constructors are not inherited. Due to historical
reasons, the Scala.js IR has supported inherited constructors, and
the compiler avoided generating so-called "trivial" constructors,
which only call their super constructor.

In this commit, we change the specification of the IR so that
constructors are not inherited anymore. For backward binary
compatibility, the analyzer still tolerates not to find a constructor
in the class where it is looked for. If it finds one in an inherited
class, the linker synthesizes the missing constructor.

As is, this introduces more code in the generated JavaScript files,
since no effort is made to avoid emitting useless JavaScript init
methods. This will be addressed in a follow up commit at the class
emitter level.

This commit does not modify the compiler yet (it still avoids
trivial constructors in the IR) so that we can test the backward
binary compatibility hack.