Give constructor methods a result type of NoType.
Previously, constructor methods explicitly returned `this`, and
advertised it in the result type of their signature. This was an
implementation detail of the encoding in JS, for which we want
init methods to return `this` so that we can chain them with `new`.

This commit fixes the specification of init methods with NoType as
result type. They do not explicitly return `this` at the IR level
anymore. It is the responsibility of the desugaring to JavaScript
to introduce the `return this`.

This commit does not change the compiler yet, so that we can test
the backward binary compatibility hack.