Update the compiler not to avoid emitting trivial constructors.
This is a follow up to the parent commit, which brings the
compiler up to date with the requirements of the IR. Since
constructors are not inherited anymore, the compiler is required
to emit all constructors, including so-called trivial constructors.

Naively, this would not change anything at this point, since
missing constructors would be recreated by the linker. However,
for module accessors with trivial constructors, the constructor
now contains a StoreModule instruction, in addition to the super
constructor call. We keep parity with the previous implementation
by adding a corresponding optimization that gets rid of unnecessary
StoreModule instructions. This optimization can be applied in a few
more cases than what the compiler did, since it also supports any
statements before the StoreModule, and any "trivial" statement after
it.