Synthesize reflective proxies at link time.
Instead of relying on reflective proxies stored in sjsir files
by the compiler, we synthesize them at link time, on demand.
During deserialization of sjsir files, we throw away all the
reflective proxies. The Analyzer identifies the reflective
proxies needed to satisfy calls of reflective methods, and the
Linker synthesizes the bridges.

In this commit, the compiler is not modified and still emits
the reflective proxies, so that we can test the backward
binary compatibility hack.