Reduce the number of classes in j.u.Collections.
Many methods of j.u.Collections wrapping collections specify that
they return a serializable collection if the wrapped collection
is serializable. The previous implementation would explicitly test
whether it was an instance of Serializable, and return different
classes inheriting from Serializable or not depending on that.

That is unnecessary. To be serializable, a class need not only be
an instance of Serializable. It must also contain only serializable
fields. Since the wrapped collection is captured by the wrapper,
it is a field, and the serializable property is carried over. We
can therefore always return collections that are instances of
Serializable, which reduces the number of classes in Collections.

This analysis is confirmed by experimentation on the JVM. Those
methods always return instances of Serializable.