Do not make ReferenceType extend Type anymore.
A ReferenceType is fundamentally *not* a Type. It so happens that
all concrete implementations of ReferenceType also implement Type,
but those two concepts must not be confused.

There are a few cases were it is OK-ish to treat a ReferenceType
as a Type (such as serializing/hashing), for which we introduced
casts.

This refactoring highlights a suspicious usage of ReferenceType as
Type in the OptimizerCore core. This commit leaves them as is,
with casts.