Remove almost all intrinsics in java.lang.Long.
The following methods of java.lang.Long were previously
intrincis, but are now left in user-space:

* bitCount
* signum
* numberOfLeadingZeros
* numberOfTrailingZeros
* toBinaryString
* toHexString
* toOctalString

With the new optimizations of `(l >>> 32).toInt` and `l.toInt`
from the parent commit, the user-space implementations can be
as efficient as their intrinsic implementation.

A few methods of java.lang.Long have been improved to take
further advantage of the efficient implementations of
`(l >>> 32).toInt` and `l.toInt`.