Deprecate bypassing linker errors.
Historically, we have allowed to bypass linking errors. This
came from a distant past where it was common for a codebase not
to link due to missing parts of the JDK. Since then, the coverage
has significantly improved, and coebases that fail to link are
not supposed to exist anymore.

When bypassing linking errors, the linking pipeline is crippled.
It will not allow the optimizer to run, for example. In the
future, it would also be very awkward to allow linker plugins to
see codebases that do not link, as their expectations might be
completely off.