Fix #2184: Reach j.l.Integer.compareTo(j.l.{Byte,Short}).
After optimizations, and in particular type refinement of
j.l.Byte into a primitive byte, which is actually an int,
it can happen that a j.l.Byte.compareTo(j.l.Byte) is
transformed into a j.l.Integer.compareTo(j.l.Byte). It is
therefore necessary to keep this method until the optimizer.
The same happens with j.l.Short.