Remove the IR node JSEnvInfo at link time.
It is now desugared as a deserialization hack into
`JSBracketSelect(JSLinkingInfo(), StringLiteral("envInfo"))`.
The `linkingInfo` object is augmented with this field `envInfo`,
pointing to the `$env` global variable.

This commit does not change the compiler and library not to emit
this node anymore, so that we can test binary backward
compatibility.