Add test detection JSON report with HTML runner.
Test detection done on sbt side is now reported via JSON to the HTML
runner. The HTML runner is therefore fully agnostic to the exact
Scala.js testing framework that is in use.
The same principle is applied in the bootstrapping test.

This fixes #2248 (Run JUnit tests in HTML test-suite runner) and takes
the first steps towards #1709 (Provide an [outside facing] HTML runner
for tests).

Not yet supported is picking up framework flags from sbt. This can be
easily added by extending the mechanism if necessary.