Fix #2358: Normalize the dir separator to '/' in relative paths.
When reading from a directory on the classpath, we need to
normalize the relative paths so that they always use '/' as
directory separator.

This commit contains no test. We already have the tests that
highlight the bug, but they only do it on Windows, and our CI
does not run on Windows.