Use the *last* From when printing analysis stack traces.
This is important because the *last* element in the list is
actually the one that was discovered *first* by the reachability
analysis. The first one is guaranteed to be on a non-cyclic path
to a FromExport or FromCore. Others are not, and can lead to
partial stack traces being displayed.