Do not display more than 20 linking errors by default.
In large projects with tens or hundreds of linking errors, the
mere fact of *printing* the linking errors takes a considerable
time (can be tens of seconds).

To prevent this, we now limit the displayed linking errors, and
only show the first 20 by default.

There should virtually never be a reason to change this limit,
but the possibility is offered through a Java system property
`org.scalajs.core.tools.linker.maxlinkingerrors`.