Fix an incorrect `null.asInstanceOf[Int]` in TreeSetTest.
`null.asInstanceOf[Int]` gives 0 under normal circumstances.
Apparently, there was a bug pre-2.12.0-M5 that would let it be
`null` instead in this particular case. But as of 2.12.0-M5,
it correctly returns 0, breaking the test.

We fix the test by doing even uglier things to force `null` to
get in the `TreeSet[Int]`.