Fix #2437: Use symbols instead of trees to list exports.
Instead of looking for `DefDef`s of exporters in a class, we look
for symbols of exporters in the class' hierarchy. This includes
superclasses and implemented interfaces. The set of found exporters
is diffed against the one we obtain for the parent class, and only
the newly found exporters are emitted.