Add PreTransUnaryOp and PreTransBinaryOp in the optimizer.
Previously, the optimizer would be stupid on the following code:

  val x: int = foo()
  val y: int = x +[int] 4
  val z: int = y *[int] 2
  // don't use any of x, y, z

When performing the `withNewLocalDef` for `y`, it would have to
reify `x` to put it in a `BinaryOp(Int_+, x, 4)`. Then, for the
`withNewLocalDef` of `z`, it would repeat the process with `y`.
Only `z` would remain virtual as a `PreTransLocalDef`.

Later on, when we discover that `z` is not used in the body of
the val, it is eliminated, and the optimizer keeps only the side
effects of its right-hand side, which collapses to `Skip()`.

However, `y` has already been marked as *used* a long time ago,
when the right hand size of `z` was first translated. It is
therefore kept, although it is clearly dead code. The same
story happens to `x`, obviously.

This commit fixes this shortcoming of the optimizer by creating
new `PreTransform`s for unary and binary ops. The rhs of `y`
is therefore kept as a virtual
`PreTransBinaryOp(Int_+, PreTransLocalDef(x), PreTransTree(4))`
which does not need to reify `x`. The same happens to `z` with
a virtual reference to `y`. After the body is translated, all
local variables can now be removed!

This only came up recently because of the sheer ability to keep
`PreTransBinding`s, which was added in
9bfc69c2a39ec3d9cab219a57c9e230cd5daa278. Now, it becomes clear
that we need `PreTransform`s for all the kinds of `Tree`s that
can be side-effect-free. This commit is a first (or second) step
towards that goal.