Fix #2466: Override checkfiles for neg/choices.scala < 2.11.8.
In scala/scala, these checkfiles have been updated in 2.11.8 and
2.12.x in the commits e606e3054a2d872135b6409f43291d5746a48cf1
and 3d7ff96bd9f637d37e9ce2922cb5140feedc8ae3 respectively.
These commits update the version of partest, so this is clearly
the source of the changes. Since we use the more recent version
of partest for all Scala versions, it makes sense that we have
to backport the changes to the checkfiles.