Fix #2453: Improve pretransformIf with one branch of type Nothing.
When one of the branches of an If is of type Nothing, we know it
cannot return normally. We can therefore rewrite the If in
statement position, and extract the other branch out of the If.
This is better because the other branch can stay fully virtual
as a PreTransform, and need not be forced to a PreTransGenTree.