Fix #2450: Locally dce VarDefs in keepOnlySideEffects.
When keeping only the side effects of a Block, we know that only
the last expression can be side-effect-free, by construction.
However, after removing the side effects of the expression, if it
collapses to Skip(), we can retry removing the side effects of
the next to last statement, in case it is a VarDef. At this point,
we know that a VarDef would be used by no expression, and it can
therefore be recursively removed.