Remove the workarounds for https://code.google.com/p/v8/issues/detail?id=3304
The v8 bug was fixed in November 2012 in
https://github.com/v8/v8/commit/23f2736a

Node.js 0.11.0 was shipped with v8 3.17.6.15
(https://github.com/v8/v8/releases/tag/3.17.6.15), which contains
that fix. We initially introduced the workaround because our CI
was running an older version of Node.js that did not have the fix,
but that time is long gone. Given the age of the fix, it is long
overdue that we remove that workaround.