When declaring a `var` of record type, use the rhs' refined type.
When declaring an immutable `val` in `withNewLocalDef`, we always
refine the type of the `val` to the `RefinedType` of the rhs.
For `var`s, in general, we cannot do that, since a later
reassignment could use a wider type.

However, when we declare the `var` as a record, we can make that
assumption. If it proves to be broken, the record will be
canceled. Previously, a reassignment to a record `var` with a
refined rhs would cancel the `var` being a record *precisely*
because the `var` had a type too wide (record assignments only
work when the refined types are exactly equal). The new way fixes
this problem.