Stack-allocate RuntimeLongs and inline their operations.
This is a major review of how the optimizer handles Longs. Longs
have been notoriously slow in Scala.js, and this commit
dramatically improves their performance (with a 15x improvement
on the SHA-512 benchmark).

The main idea is two-fold:

* Mark `RuntimeLong` with `@inline`
* Have the optimizer eagerly expand all `Long` operations into the
  method calls that implement them.

This combination makes it so that `RuntimeLong`s are virtualized
and stack-allocated, and that all their operations are inlined and
work on the `Int` values `lo` and `hi`. We also accompany those
two core changes with a few tricks--such as allocating new
instances of `RuntimeLong` on escape, rather than canceling, which
we can do because the identity of `RuntimeLong`s is not
meaningful--to push the amount of stack allocation as much as
possible.

This has two advantages. The first and foremost is that it is not
necessary to allocate a new instance of `RuntimeLong` on heap for
every single operation. Instead, instances are only created when
the value escapes the scope of the current method.

The second advantage is that the optimizer can micro-optimize the
operations specifically for their call site. For example, if one
operand is a constant, as in `1L & x`, the lo and hi parts of the
constant are actually constant themselves, and will be treated as
such while optimizing "the inside" of the `&` operation. Which
gives something like:

  val result$lo = 1 & x$lo;
  val result$hi = 0 & x$hi;

the latter then reduces to `0`, which in turns causes `result$hi`
to be constant for subsequent operations.

Here are some of my favorite rewritings.

* `(x >>> 32).toInt` becomes `x$hi` (this was previously
  special-cased in the optimizer, but is now a consequence of the
  new optimizations)
* `x >= 0L` becomes `x$hi >= 0`, and `x < 0L` becomes `x$hi < 0`
* `(xInt.toLong << 32) | (yInt.toLong & 0xffffffffL)` becomes
  `val result$lo = xInt; val result$hi = yInt;`
* Even more impressive,
  `(xInt.toLong << 32) + (yInt.toLong & 0xffffffffL)` (with a `+`
  instead of `|`) also reduces to
  `val result$lo = xInt; val result$hi = yInt;`
  (but if the operands are swapped, that doesn't work so well
  anymore).
* `(xInt.toLong & 0xffffffffL) > (yInt.toLong & 0xffffffffL)`
  rewrites as `(xInt ^ 0x80000000L) > (yInt ^ 0x80000000L)`, and
  similarly for the other comparison operators (note that this is
  the "optimized by hand" version of comparing `xInt` and `yInt`
  as if they were unsigned integers)